display bonuses burning date

+++ bonusesview/src/main/java/com/grizzlyorange/bonusesview/data/BonusesViewInfo.kt
@@ -3,5 +3,5 @@ package com.grizzlyorange.bonusesview.data
 data class BonusesViewInfo (
     val totalBonusesAmount: Int = 0,
     val burningBonusesAmount: Int = 0,
-    val burningDateInMillis: Int = 0
+    val burningDateInMillis: Long = 0
 )
